requirements.txt
beautifulsoup4==4.13.5
requests==2.32.5
numpy==2.2.6
pandas==2.3.2
openpyxl==3.1.5
fastapi==0.116.2
uvicorn==0.35.0
pydantic==2.11.9
google-generativeai==0.8.5